+++ Amplitude/AMPConstants.h
@@ -16,3 +16,10 @@ extern const int kAMPEventMaxCount;
 extern const int kAMPEventRemoveBatchSize;
 extern const int kAMPEventUploadPeriodSeconds;
 extern const long kAMPMinTimeBetweenSessionsMillis;
+extern const int kAMPMaxStringLength;
+
+extern NSString *const IDENTIFY_EVENT;
+extern NSString *const AMP_OP_ADD;
+extern NSString *const AMP_OP_SET;
+extern NSString *const AMP_OP_SET_ONCE;
+extern NSString *const AMP_OP_UNSET;