+++ AmplitudeTests/SetupTests.m
@@ -12,6 +12,7 @@
 #import "Amplitude.h"
 #import "Amplitude+Test.h"
 #import "BaseTestCase.h"
+#import "AMPConstants.h"
 #import "AMPUtils.h"
 
 @interface SetupTests : BaseTestCase
@@ -78,6 +79,8 @@
 
 - (void)testUserPropertiesSet {
     [self.amplitude initializeApiKey:apiKey];
+    AMPDatabaseHelper *dbHelper = [AMPDatabaseHelper getDatabaseHelper];
+    XCTAssertEqual([dbHelper getEventCount], 0);
 
     NSDictionary *properties = @{
          @"shoeSize": @10,
@@ -85,42 +88,19 @@
          @"name": @"John"
     };
 
-    [self.amplitude setUserProperties:@{@"property": @"true"} replace:YES];
-    [self.amplitude setUserProperties:properties replace:YES];
-
-    [self.amplitude logEvent:@"Test Event"];
-    [self.amplitude flushQueue];
-
-    NSDictionary *event = [self.amplitude getLastEvent];
-    XCTAssert([event[@"user_properties"] isEqualToDictionary:properties]);
-}
-
-- (void)testUserPropertiesMerge {
-    [self.amplitude initializeApiKey:apiKey];
-
-    NSMutableDictionary *properties = [@{
-         @"shoeSize": @10,
-         @"hatSize":  @5.125,
-         @"name": @"John"
-    } mutableCopy];
-
     [self.amplitude setUserProperties:properties];
-
-    [self.amplitude logEvent:@"Test Event"];
     [self.amplitude flushQueue];
+    XCTAssertEqual([dbHelper getEventCount], 0);
+    XCTAssertEqual([dbHelper getIdentifyCount], 1);
+    XCTAssertEqual([dbHelper getTotalEventCount], 1);
 
-    NSDictionary *event = [self.amplitude getLastEvent];
-    XCTAssert([event[@"user_properties"] isEqualToDictionary:properties]);
+    NSDictionary *expected = [NSDictionary dictionaryWithObject:properties forKey:AMP_OP_SET];
 
-    NSDictionary *extraProperties = @{@"mergedProperty": @"merged"};
-    [self.amplitude setUserProperties:extraProperties replace:NO];
-
-    [self.amplitude logEvent:@"Test Event"];
-    [self.amplitude flushQueue];
-
-    event = [self.amplitude getLastEvent];
-    [properties addEntriesFromDictionary:extraProperties];
-    XCTAssert([event[@"user_properties"] isEqualToDictionary:properties]);
+    NSDictionary *event = [self.amplitude getLastIdentify];
+    XCTAssertEqualObjects([event objectForKey:@"event_type"], IDENTIFY_EVENT);
+    XCTAssertEqualObjects([event objectForKey:@"user_properties"], expected);
+    XCTAssertEqualObjects([event objectForKey:@"event_properties"], [NSDictionary dictionary]); // event properties should be empty
+    XCTAssertEqual(1, [[event objectForKey:@"sequence_number"] intValue]);
 }
 
 - (void)testSetDeviceId {